+++ cmd2/exceptions.py
@@ -62,6 +62,21 @@ class CompletionError(Exception):
         super().__init__(*args)
 
 
+class PassThroughException(Exception):
+    """
+    Normally all unhandled exceptions raised during commands get printed to the user.
+    This class is used to wrap an exception that should be raised instead of printed.
+    """
+
+    def __init__(self, *args, wrapped_ex: BaseException):
+        """
+        Initializer for PassThroughException
+        :param wrapped_ex: the exception that will be raised
+        """
+        self.wrapped_ex = wrapped_ex
+        super().__init__(*args)
+
+
 ############################################################################################################
 # The following exceptions are NOT part of the public API and are intended for internal use only.
 ############################################################################################################